Implement sheetStore.setDefaultData() on the frontend
On the front-end JSON.parse(sheetData) is in a try catch block. If anything went
wrong there, the sheetstore will revert to default data. To link it up, I
implemented a lazy instantiation of the main.sections array in
sheetDataManager.addSection(). This way a section can always be creating,
whichever the state of the data is. Wrote specs.