Change lastRef in sheetStore
Was: {chordID: .., parentIDs: ..}
Is now: {chordID: .., barID: .., rowID: .., sectionID: ..}
Easier to manage missing params. Example:
if the sheet is empty and has just been opened, there is no
'lastSelectedChordRef'. adding a bar with nested params it will throw when trying to access
ref.parentIDs.barID, but flat it will just send along undefined, which will then
be interpreted as 'oh just put it in the sheet wherever'