Update sheet.findById to throw custom error if not found. Update factory
The sheet class will throw a "Sheet not found" error, which can in turn be
caught in the controllers/routes to handle different results.
The factory had a params bug where the params in the if state in the switch case
would suddenly be undefined (??) I had to hack and bind the object from the
outside. This is really weird!