Implement dynamic friendship label responses. Huge refactoring for front-back communication
The labels now display the friendship status in 'real-time'. Implemented a
backbone model to represent the friendship connected to the friends controller,
and one to represent the friendrequests connected to the friendrequests
controller. These are nested models in the user store and sync up when
appropriate, triggering the events caught by the main userpage react component.

-- This does however result in one too many ajax requests: one when clicking a
friendship-related button, and one after the friendship has been updated to get
the new state. I might refactor to try to do all in a single request somehow.