Front-end permission selector sends update req to back-end + gets auto set to current state
Setup the floxbone route from the change event on the permissions selector in
the participant manager modal to send a PUT request to
/hubs/:hub_id/invitations/:invitation_id with the slider's current value.
When invitations are fetched, the slider's also are set to represent the
permission current invitation permission state.
Also cleared up some remaining console.log statements all over the front-end.