Implement default permissions storage upon sending an invitation controller+front-end
For now a simple DEFAULT_PERMISSIONS var is stored in the hubStore. Later this
value will be pulled from the hub's settings.
When sending an invitation the default permissions val is sent along with the
invitation. Implement the POST/create controller for hub_invitations to take
into account this extra parameter.