Implement confirmation modal dropdown for removing participants from hub.
I had to modify the modal slightly. I have added some dynamism to the modal
react class. We now have the options to set the display text and confirmation
callback at runtime.

Added a flux route for showing the modal. When sent, the store emits a
modal:confirm event with the modal params, the main component then dynamically
sets the data for the confirmation modal and slides it out.