Wrote specs for a bug in user.acceptHubInvitation
It was a typo bug, but here's what happened:
When userA had created a hub and sent invitations to userB and userC, when userC
accepted the hubInvitation, it would create TWO JOINED relationships between
userC and the hub, and not delete the other hubInvitation, but delete the
TO_JOIN relationship between the other hubInvitation and the hub.
Turns out I had a var named 'hi' and 'invitation' in my cypher query, but I
meant both to reference the same HubInvitation node. Wrote two extra specs to
prevent this from happening again.