Major sheetEditor refactoring!! -> migrated store to be a backbone model.
The sheetStore is now a backbone model, following the fluxBone pattern. I have
moved the dispatch callback into the store (instead of in the dispatcher file).
I have merged the network actions and sheetactions to be just one actions file
(because backbone manages it's on networking). I have changed the instantiation
of the sheet editor react component to just take in the store as prop.